Comms console in TS (#89106) Refactors the communications console into typescript. Removed a uselocalstate. I tried not to dig too deep, this is an ugly UI tbh. I cleaned up some of the logic (the changing alert level state was weird) and broke it up into files / components where possible.
Style wise, only the purchase shuttle screen looks slightly different,
with the "back" button and the budget now occupying the same section.
Code quality
N/A
