+++ tgui/packages/tgui/interfaces/CommunicationsConsole/Main.tsx
@@ -1,0 +1,331 @@
+import { useState } from 'react';
+import { Box, Button, Flex, Modal, Section } from 'tgui-core/components';
+import { capitalize } from 'tgui-core/string';
+
+import { useBackend } from '../../backend';
+import { AlertButton } from './AlertButton';
+import { MessageModal } from './MessageModal';
+import { CommsConsoleData, ShuttleState } from './types';
+
+export function PageMain(props) {
+  const { act, data } = useBackend<CommsConsoleData>();
+  const {
+    alertLevel,
+    callShuttleReasonMinLength,
+    canBuyShuttles,
+    canMakeAnnouncement,
+    canMessageAssociates,
+    canRecallShuttles,
+    canRequestNuke,
+    canSendToSectors,
+    canSetAlertLevel,
+    canToggleEmergencyAccess,
+    emagged,
+    syndicate,
+    emergencyAccess,
+    importantActionReady,
+    sectors,
+    shuttleCalled,
+    shuttleCalledPreviously,
+    shuttleCanEvacOrFailReason,
+    shuttleLastCalled,
+    shuttleRecallable,
+  } = data;
+
+  const [callingShuttle, setCallingShuttle] = useState(false);
+  const [messagingAssociates, setMessagingAssociates] = useState(false);
+  const [messagingSector, setMessagingSector] = useState('');
+  const [requestingNukeCodes, setRequestingNukeCodes] = useState(false);
+
+  const [newAlertLevel, setNewAlertLevel] = useState('');
+  const showAlertLevelConfirm = newAlertLevel && newAlertLevel !== alertLevel;
+
+  return (
+    <Box>
+      {!syndicate && (
+        <Section title="Emergency Shuttle">
+          {shuttleCalled ? (
+            <Button.Confirm
+              icon="space-shuttle"
+              color="bad"
+              disabled={!canRecallShuttles || !shuttleRecallable}
+              tooltip={
+                (canRecallShuttles &&
+                  !shuttleRecallable &&
+                  "It's too late for the emergency shuttle to be recalled.") ||
+                'You do not have permission to recall the emergency shuttle.'
+              }
+              tooltipPosition="top"
+              onClick={() => act('recallShuttle')}
+            >
+              Recall Emergency Shuttle
+            </Button.Confirm>
+          ) : (
+            <Button
+              icon="space-shuttle"
+              disabled={shuttleCanEvacOrFailReason !== 1}
+              tooltip={
+                shuttleCanEvacOrFailReason !== 1
+                  ? shuttleCanEvacOrFailReason
+                  : undefined
+              }
+              tooltipPosition="top"
+              onClick={() => setCallingShuttle(true)}
+            >
+              Call Emergency Shuttle
+            </Button>
+          )}
+          {!!shuttleCalledPreviously &&
+            (shuttleLastCalled ? (
+              <Box>
+                Most recent shuttle call/recall traced to:{' '}
+                <b>{shuttleLastCalled}</b>
+              </Box>
+            ) : (
+              <Box>Unable to trace most recent shuttle/recall signal.</Box>
+            ))}
+        </Section>
+      )}
+
+      {!!canSetAlertLevel && (
+        <Section title="Alert Level">
+          <Flex justify="space-between">
+            <Flex.Item>
+              <Box>
+                Currently on <b>{capitalize(alertLevel)}</b> Alert
+              </Box>
+            </Flex.Item>
+
+            <Flex.Item>
+              <AlertButton
+                alertLevel="green"
+                onClick={() => setNewAlertLevel('green')}
+              />
+
+              <AlertButton
+                alertLevel="blue"
+                onClick={() => setNewAlertLevel('blue')}
+              />
+            </Flex.Item>
+          </Flex>
+        </Section>
+      )}
+
+      <Section title="Functions">
+        <Flex direction="column">
+          {!!canMakeAnnouncement && (
+            <Button
+              icon="bullhorn"
+              onClick={() => act('makePriorityAnnouncement')}
+            >
+              Make Priority Announcement
+            </Button>
+          )}
+
+          {!!canToggleEmergencyAccess && (
+            <Button.Confirm
+              icon="id-card-o"
+              confirmIcon="id-card-o"
+              color={emergencyAccess ? 'bad' : undefined}
+              onClick={() => act('toggleEmergencyAccess')}
+            >
+              {emergencyAccess ? 'Disable' : 'Enable'} Emergency Maintenance
+              Access
+            </Button.Confirm>
+          )}
+
+          {!syndicate && (
+            <Button
+              icon="desktop"
+              onClick={() =>
+                act('setState', { state: ShuttleState.CHANGING_STATUS })
+              }
+            >
+              Set Status Display
+            </Button>
+          )}
+
+          <Button
+            icon="envelope-o"
+            onClick={() => act('setState', { state: ShuttleState.MESSAGES })}
+          >
+            Message List
+          </Button>
+
+          {canBuyShuttles !== 0 && (
+            <Button
+              icon="shopping-cart"
+              disabled={canBuyShuttles !== 1}
+              // canBuyShuttles is a string detailing the fail reason
+              // if one can be given
+              tooltip={canBuyShuttles !== 1 ? canBuyShuttles : undefined}
+              tooltipPosition="top"
+              onClick={() =>
+                act('setState', { state: ShuttleState.BUYING_SHUTTLE })
+              }
+            >
+              Purchase Shuttle
+            </Button>
+          )}
+
+          {!!canMessageAssociates && (
+            <Button
+              icon="comment-o"
+              disabled={!importantActionReady}
+              onClick={() => setMessagingAssociates(true)}
+            >
+              Send message to {emagged ? '[UNKNOWN]' : 'CentCom'}
+            </Button>
+          )}
+
+          {!!canRequestNuke && (
+            <Button
+              icon="radiation"
+              disabled={!importantActionReady}
+              onClick={() => setRequestingNukeCodes(true)}
+            >
+              Request Nuclear Authentication Codes
+            </Button>
+          )}
+
+          {!!emagged && !syndicate && (
+            <Button icon="undo" onClick={() => act('restoreBackupRoutingData')}>
+              Restore Backup Routing Data
+            </Button>
+          )}
+        </Flex>
+      </Section>
+
+      {!!canMessageAssociates && messagingAssociates && (
+        <MessageModal
+          label={`Message to transmit to ${
+            emagged ? '[ABNORMAL ROUTING COORDINATES]' : 'CentCom'
+          } via quantum entanglement`}
+          notice="Please be aware that this process is very expensive, and abuse will lead to...termination. Transmission does not guarantee a response."
+          icon="bullhorn"
+          buttonText="Send"
+          onBack={() => setMessagingAssociates(false)}
+          onSubmit={(message) => {
+            setMessagingAssociates(false);
+            act('messageAssociates', {
+              message,
+            });
+          }}
+        />
+      )}
+
+      {!!canRequestNuke && requestingNukeCodes && (
+        <MessageModal
+          label="Reason for requesting nuclear self-destruct codes"
+          notice="Misuse of the nuclear request system will not be tolerated under any circumstances. Transmission does not guarantee a response."
+          icon="bomb"
+          buttonText="Request Codes"
+          onBack={() => setRequestingNukeCodes(false)}
+          onSubmit={(reason) => {
+            setRequestingNukeCodes(false);
+            act('requestNukeCodes', {
+              reason,
+            });
+          }}
+        />
+      )}
+
+      {!!callingShuttle && (
+        <MessageModal
+          label="Nature of emergency"
+          icon="space-shuttle"
+          buttonText="Call Shuttle"
+          minLength={callShuttleReasonMinLength}
+          onBack={() => setCallingShuttle(false)}
+          onSubmit={(reason) => {
+            setCallingShuttle(false);
+            act('callShuttle', {
+              reason,
+            });
+          }}
+        />
+      )}
+
+      {!!canSetAlertLevel && showAlertLevelConfirm && (
+        <Modal>
+          <Flex direction="column" textAlign="center" width="300px">
+            <Flex.Item fontSize="16px" mb={2}>
+              Swipe ID to confirm change
+            </Flex.Item>
+
+            <Flex.Item mr={2} mb={1}>
+              <Button
+                icon="id-card-o"
+                color="good"
+                fontSize="16px"
+                onClick={() => {
+                  act('changeSecurityLevel', {
+                    newSecurityLevel: newAlertLevel,
+                  });
+                  setNewAlertLevel('');
+                }}
+              >
+                Swipe ID
+              </Button>
+
+              <Button
+                icon="times"
+                color="bad"
+                fontSize="16px"
+                onClick={() => setNewAlertLevel('')}
+              >
+                Cancel
+              </Button>
+            </Flex.Item>
+          </Flex>
+        </Modal>
+      )}
+
+      {!!canSendToSectors && sectors.length > 0 && (
+        <Section title="Allied Sectors">
+          <Flex direction="column">
+            {sectors.map((sectorName) => (
+              <Flex.Item key={sectorName}>
+                <Button
+                  disabled={!importantActionReady}
+                  onClick={() => setMessagingSector(sectorName)}
+                >
+                  Send a message to station in {sectorName} sector
+                </Button>
+              </Flex.Item>
+            ))}
+
+            {sectors.length > 2 && (
+              <Flex.Item>
+                <Button
+                  disabled={!importantActionReady}
+                  onClick={() => setMessagingSector('all')}
+                >
+                  Send a message to all allied station
+                </Button>
+              </Flex.Item>
+            )}
+          </Flex>
+        </Section>
+      )}
+
+      {!!canSendToSectors && sectors.length > 0 && messagingSector && (
+        <MessageModal
+          label="Message to send to allied station"
+          notice="Please be aware that this process is very expensive, and abuse will lead to...termination."
+          icon="bullhorn"
+          buttonText="Send"
+          onBack={() => setMessagingSector('')}
+          onSubmit={(message) => {
+            act('sendToOtherSector', {
+              destination: messagingSector,
+              message,
+            });
+
+            setMessagingSector('');
+          }}
+        />
+      )}
+    </Box>
+  );
+}
