+++ tgui/packages/tgui/interfaces/CommunicationsConsole/NoConnectionModal.tsx
@@ -1,0 +1,31 @@
+import { Blink, Dimmer, Flex, Icon } from 'tgui-core/components';
+
+export function NoConnectionModal() {
+  return (
+    <Dimmer>
+      <Flex direction="column" textAlign="center" width="300px">
+        <Flex.Item>
+          <Icon color="red" name="wifi" size={10} />
+
+          <Blink>
+            <div
+              style={{
+                background: '#db2828',
+                bottom: '60%',
+                left: '25%',
+                height: '10px',
+                position: 'relative',
+                transform: 'rotate(45deg)',
+                width: '150px',
+              }}
+            />
+          </Blink>
+        </Flex.Item>
+
+        <Flex.Item fontSize="16px">
+          A connection to the station cannot be established.
+        </Flex.Item>
+      </Flex>
+    </Dimmer>
+  );
+}
