+++ tgui/packages/tgui/interfaces/CommunicationsConsole/constants.ts
@@ -1,0 +1,5 @@
+// Used for whether or not you need to swipe to confirm an alert level change
+export const SWIPE_NEEDED = 'SWIPE_NEEDED';
+
+export const EMAG_SHUTTLE_NOTICE =
+  'This shuttle is deemed significantly dangerous to the crew, and is only supplied by the Syndicate.';
